do not take embedded codes as folding elements
currently MarkdownLevelSA() does not check that lines starting with
hashes are in an embedded code block. It means that if i have line

```vim
let g:pandoc#syntax#codeblocks#embeds#langs = ["cpp"]
```

in my .vimrc and a document containing cpp macros like

```c
 #include <string.h>
```

then they will be erroneously taken as folding elements.

This fix checks that synId is not empty and does not start with 'pandoc'
and additionaly that ``g:pandoc#syntax#codeblocks#embeds#langs`` exists
and not empty and in this case MarkdownLevelSA() will return '=' which
will prevent presumably embedded code block from interpreting as a new
folding element.

BEWARE! This fix also requires changes in pandoc syntax file. AtxStart
which (hopefully accidentally) is the only pandoc syntax element that
does not start with 'pandoc' must be renamed to 'pandocAtxStart'.