bibliography, completion: mayor restructuring of the bibliographic completion
started work on integrating citeproc-based completion.

introduced new config, g:pandoc#completion#bib#mode, that sets which
backend to use for information retrieval.

the new entry point is vim_pandoc.bib.vim_completer, which is the only
point where the completion system interacts with vim.

split SourceCollator from citeproc.py, and made it the root class of
completion backends. it handles the location of bibliographic databases.
The SourceCollator API has changed, queries are given in initialization
instead of on collate().

turned the old mechanism (fallback) into an subclass of the new collator
system. it is used by default while the citeproc-based mechanism is
being worked on.

MODS support removed from fallback system (it was buggy).