Apply vint to autoload
add `abort` to autoload functinos

Uses
[vint](https://github.com/Kuniwak/vint/wiki/Vint-linting-policy-summary)
suggestions.

Explanation of `abort` from vint: [`abort` forces
the function to halt when it encounters an error.](
https://google.github.io/styleguide/vimscriptguide.xml?showone=Functions

Make boolean comparison operators more robust

`==`, `=~`, and `!=`, and `!=~` depend on the users case sensitivity
settings.

You need to add `#` or `?` at the end of them to make them explicitly
case sensitive or insensitive.