Adds support for Pandoc 1.19 and above.
In Pandoc 1.18 the help syntax to find input and output formats was
changed. To that end there was made an exception that checked
specifically for a version that was 1.18 in order to find the formats in
this new version. Pandoc is now in version 1.19 and follows the same
pattern as 1.18 and it is expected to continue like that, so as an easy
fix the if condition has been changed to include future versions as
well.