Fixed bug with `j` and `k` user map restoration
In issue #198, some code was added to formatting.vim that is supposed to
restore user mappings for j and k, if they existed beforehand. The
normal mode mapping was being restored correctly, but the visual mode
mapping was still getting dropped on the floor. It turns out the blocks
for the visual and normal mappings were exact copies of each other. This
commit makes the minor tweak necessary to actually restore the visual
mode mappings.