Remove chdir module
As discussed at https://github.com/vim-pandoc/vim-pandoc/issues/272 the chdir
module, previously being enabled by default, was removed. Please add
>
  set autochdir
<
to your vimrc in case you want the current work directory of every window to
be set to that of the file opened in the buffer. See also *autochdir*.